美化06_utilities.c文件：改进变量命名规范和函数文档注释
- 重命名Unwind_18090ad90为CleanupObjectContextResources，添加详细的函数文档
- 重命名Unwind_18090adb0为ValidateObjectContextIntegrity，添加安全验证说明
- 重命名Unwind_18090add0为ResetObjectContextState，添加状态重置说明
- 重命名Unwind_18090adf0为InitializeObjectContext，添加初始化说明
- 重命名Unwind_18090ae30为FreeObjectContextMemory，添加内存管理说明
- 重命名Unwind_18090ae70为ValidateObjectContextSecurity，添加安全验证说明
- 统一使用PascalCase命名规范，提高代码可读性
- 为所有函数添加详细的中文文档注释，包含参数说明和使用注意事项

+++ TaleWorlds.Native/src/01_initialization.c
@@ -19708,7 +19708,7 @@ void ProcessSystemThreeParameterBuffer(long long MainParameter,long long Auxilia
       memcpy(SystemBufferPointer,*(long long *)(MainParameter + 8),StringSearchResult - *(long long *)(MainParameter + 8));
   }
   SystemMemoryReference = &SystemMemoryAllocatorReference;
-    ValidateSystemChecksum(SystemChecksumValue ^ (ulong long)StackSecurityBuffer);
+    ValidateSystemChecksum(SystemChecksumValue ^ (unsigned long long)StackSecurityBuffer);
 }
